requirements.txt
Django>=5.2,<6.0
# REST framework
djangorestframework>=3.14,<4.0
# JWT
djangorestframework-simplejwt>=5.3,<6.0
# CORS
django-cors-headers>=4.3,<5.0
# Env loading
python-dotenv>=1.0,<2.0
# Images
Pillow>=10.0,<11.0
# WebAuthn
webauthn>=2.0,<3.0
# Optional browser automation used by dynamic barcode code (safe even if disabled)
selenium>=4.0,<5.0
# Useful dev utilities referenced in settings
django-extensions>=3.2,<4.0
django-widget-tweaks>=1.5,<2.0
# WSGI server
gunicorn>=21.2,<22.0
# Optional Postgres driver (not used when defaulting to sqlite3)
psycopg2-binary>=2.9,<3.0


